+++ src/mappers/bots/dca/map-bot-settings-to-form-data.ts
@@ -869,7 +869,7 @@ export const mapBotSettingsToFormData = (
     comboSlLimit: getBoolean('comboSlLimit', false),
     useSmartOrders: getBoolean('useSmartOrders'),
     dcaByMarket: getBoolean('dcaByMarket', false),
-    rejectBelowExchangeMin: getBoolean('rejectBelowExchangeMin', false),
+    allowRaiseToExchangeMin: getBoolean('allowRaiseToExchangeMin', false),
     maxNumberOfOpenDeals: getString('maxNumberOfOpenDeals', '1'),
     type: getValue<DCATypeEnum>('type', DCATypeEnum.regular),
     startDealLogic,